Work in progress.

+++ aeeg_dialog.h
@@ -39,8 +39,8 @@
 #include "mainwindow.h"
 #include "utils.h"
 #include "filteredblockread.h"
-//#include "aeeg_dock.h"
-
+#include "aeeg_dock.h"
+#include "third_party/fidlib/fidlib.h"
 
 
 class UI_Mainwindow;
@@ -82,8 +82,8 @@ private slots:
   void start_button_clicked();
   void default_button_clicked();
   void segmentlen_spinbox_changed(int);
-  void min_hz_spinbox_changed(int);
-  void max_hz_spinbox_changed(int);
+  void min_hz_spinbox_changed(double);
+  void max_hz_spinbox_changed(double);
 
 };
